% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{blockShiny}
\alias{blockShiny}
\title{blockShiny}
\usage{
blockShiny(shiny.maxRequestSize = 100 * 1024^2, launch.browser = T)
}
\arguments{
\item{shiny.maxRequestSize}{The maximum size of the uploaded file.}

\item{launch.browser}{If `TRUE`, launch browser.}
}
\description{
The function \code{blockShiny} starts the browser to start anlysis.
}
\examples{
library (metaFunc)
blockShiny()

}
