% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockplot.R
\name{blockPlot}
\alias{blockPlot}
\title{blockPlot}
\usage{
blockPlot(func_data, tax_data, func_split = ";", split_percentage = 10,
show_func_topn = 20, rownames_size = 10, tax_name_size = 3,
legend_label_size = 10, avoid_label_overlap = T)
}
\arguments{
\item{func_data}{functional annotation data}

\item{tax_data}{tax annotation data}

\item{func_split}{split}

\item{show_func_topn}{show top n of functioanl term}

\item{split_percentage}{split percentage}

\item{rownames_size}{the size of rownames}

\item{tax_name_size}{the size of tax name}

\item{legend_label_size}{the size of legend size}

\item{avoid_label_overlap}{avoid label overlapping}
}
\description{
The function \code{blockPlot} starts the browser to start anlysis.
}
\examples{
library (metaFunc)
data(simple_demo)
blockPlot(func_data = simple_demo$func, tax_data = simple_demo$tax)

}
