% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{blockShiny}
\alias{blockShiny}
\title{blockShiny}
\usage{
blockShiny()
}
\arguments{
\item{shiny.maxRequestSize}{max upload size of file}

\item{launch.browser}{open in browser}
}
\description{
The function \code{blockShiny} starts the browser to start anlysis.
}
\examples{
library (metaFunc)
blockShiny()

}
